[lldb] Step over non-lldb breakpoints
Several languages support some sort of "breakpoint" intrinsic, which
adds ISA-specific instructions to generate an interrupt at runtime.
However, on some platforms, these instructions don't increment the
program counter. When LLDB sets these instructions it isn't a problem,
as we remove them before continuing, then re-add them after stepping
over the location. However, for breakpoint sequences that are in the
inferior process, this doesn't happen - and so users might be left
unable to continue past the breakpoint without manually interfering
with the program counter.

This patch adds logic to LLDB to intercept SIGTRAPs, inspect the bytes
of the inferior at the program counter, and if the instruction looks
like a BRK or BKPT or similar, increment the pc by the size of the
instruction we found. This unifies platform behaviour (i.e. on x86_64,
LLDB debug sessions already look like this) and improves UX (in my
opinion, but I think it beats messing with stuff every break).